requirements.txt
# Requirements for evaluation app

django-admin-bootstrapped
django-report-builder
django-reversion-compare
    # after every model change, run:
    # python manage.py createinitialrevisions